third_party/blink/web_tests/fast/css/font-shorthand-from-longhands-expected.txt

Test the return values for the font properties on the style object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.font is ''
PASS computedStyle.font is '20px foobar'
PASS computedStyle.fontSize is '20px'
PASS checkFontStyleValue() is true
PASS style.font is ''
PASS computedStyle.font is '20px sans-serif'
PASS computedStyle.fontFamily is 'sans-serif'
PASS style.font is ''
PASS computedStyle.font is 'italic 20px sans-serif'
PASS computedStyle.fontStyle is 'italic'
PASS style.font is ''
PASS computedStyle.font is 'italic small-caps 20px sans-serif'
PASS computedStyle.fontVariant is 'small-caps'
PASS style.font is ''
PASS computedStyle.font is 'italic small-caps 700 20px sans-serif'
PASS computedStyle.fontWeight is '700'
PASS style.font is ''
PASS computedStyle.font is 'italic small-caps 700 20px / 40px sans-serif'
PASS computedStyle.lineHeight is '40px'
PASS style.font is 'italic small-caps bold ultra-expanded 20px / 40px sans-serif'
PASS computedStyle.font is 'italic small-caps 700 ultra-expanded 20px / 40px sans-serif'
PASS checkFontStyleValue() is true
PASS style.font is ''
PASS computedStyle.font is ''
PASS style.font is 'italic small-caps bold ultra-expanded 20px / 40px sans-serif'
PASS computedStyle.font is 'italic small-caps 700 ultra-expanded 20px / 40px sans-serif'
PASS style.font is ''
PASS computedStyle.font is ''
PASS style.font is 'italic small-caps bold ultra-expanded 20px / 40px sans-serif'
PASS computedStyle.font is 'italic small-caps 700 ultra-expanded 20px / 40px sans-serif'
PASS style.font is ''
PASS computedStyle.font is '16px foobar'
PASS checkFontStyleValue() is true
PASS style.fontVariantCaps is 'all-small-caps'
PASS computedStyle.font is ''
PASS style.font is ''
PASS successfullyParsed is true

TEST COMPLETE
